Crate icu_compactdecimal

🚧 [Experimental] Relative time formatting

This module is published as its own crate (icu_compactdecimal) and as part of the icu crate. See the latter for more details on the ICU4X project.

🚧 This code is experimental; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. It can be enabled with the "experimental" feature of the icu meta-crate. Use with caution.

Modules

provider

Data provider struct definitions for this ICU4X component.

Structs

CompactDecimalFormatter

A formatter that renders locale-sensitive compact numbers.

Enums

CompactDecimalError

A list of error outcomes for various operations in the icu_compactdecimal crate.