Module icu_compactdecimal::provider

Data provider struct definitions for this ICU4X component.

Read more about data providers: icu_provider

Structs

CompactDecimalPatternDataV1

Relative time format V1 data struct. As in CLDR, this is a mapping from type (a power of ten, corresponding to the magnitude of the number being formatted) and count (a plural case or an explicit 1) to a pattern.

CountULE

ULE type for Count

LongCompactDecimalFormatDataV1Marker

Marker type for CompactDecimalPatternDataV1: “compactdecimal/long@1”

Pattern

A compact decimal pattern, representing some literal text with an optional placeholder, and the power of 10 expressed by the text.

PatternULE

VarULE type for Pattern

ShortCompactDecimalFormatDataV1Marker

Marker type for CompactDecimalPatternDataV1: “compactdecimal/short@1”

Enums

Count

A CLDR plural keyword, or the explicit value 1. See https://www.unicode.org/reports/tr35/tr35-numbers.html#Language_Plural_Rules.